refactor(elasticsearch): split backfill/incremental/cdc into separate files
