+++ drivers/elasticsearch/internal/config.go
@@ -37,10 +37,9 @@ func (c *Config) Validate() error {
 		}
 	}
 
-	// Authentication is optional for testing
-	// if c.APIKey == "" && c.Username == "" {
-	//	return fmt.Errorf("either api_key or username/password must be provided")
-	// }
+	if c.CloudID == "" && c.APIKey == "" && c.Username == "" {
+		return fmt.Errorf("either api_key or username/password must be provided")
+	}
 
 	if c.Username != "" && c.Password == "" {
 		return fmt.Errorf("password is required when username is provided")
